requirements.txt
black           # PEP8 formatter
ipython         # great Python interpreter
nicegui         # cool Python GUI
notebook        # all Jupyter notebook needs
pipreqs         # simple requirements
plotly          # interactive plots (with JS)
pylint          # great linter
pytest          # unit tests
pytest-cov      # unit tests, cover a specific functionpytest-cov
pyyaml          # YAML support
seaborn         # plots
streamlit       # cool Python GUI
